Data.Bifunctor.Joker

Description

From the Functional Pearl "Clowns to the Left of me, Jokers to the Right: Dissecting Data Structures" by Conor McBride.

Documentation

newtype Joker g a b

Make a Functor over the first argument of a Bifunctor.

Constructors

Joker

Fields

runJoker :: g b