Data.Biapplicative

Description

Biapplicative bifunctors

class Bifunctor p => Biapplicative p where

Minimal complete definition

bipure, ((<<*>>) | biliftA2)

Methods

bipure :: a -> b -> p a b

(<<*>>) :: p (a -> b) (c -> d) -> p a c -> p b d infixl 4

biliftA2 :: (a -> b -> c) -> (d -> e -> f) -> p a d -> p b e -> p c f

Lift binary functions

(*>>) :: p a b -> p c d -> p c d infixl 4

a *>> b ≡ bimap (const id) (const id) <<$>> a <<*>> b

(<<*) :: p a b -> p c d -> p a b infixl 4

a <<* b ≡ bimap const const <<$>> a <<*>> b

(<<$>>) :: (a -> b) -> a -> b infixl 4

(<<**>>) :: Biapplicative p => p a c -> p (a -> b) (c -> d) -> p b d infixl 4

biliftA3 :: Biapplicative w => (a -> b -> c -> d) -> (e -> f -> g -> h) -> w a e -> w b f -> w c g -> w d h

Lift ternary functions

traverseBia :: (Traversable t, Biapplicative p) => (a -> p b c) -> t a -> p (t b) (t c)

Traverse a Traversable container in a Biapplicative.

traverseBia satisfies the following properties:

Pairing

traverseBia (,) t = (t, t)

Composition

traverseBia (Biff . bimap g h . f) = Biff . bimap (traverse g) (traverse h) . traverseBia f

traverseBia (Tannen . fmap f . g) = Tannen . fmap (traverseBia f) . traverse g

Naturality

 t . traverseBia f = traverseBia (t . f)

for every biapplicative transformation t.

A biapplicative transformation from a Biapplicative P to a Biapplicative Q is a function

t :: P a b -> Q a b

preserving the Biapplicative operations. That is,

t (bipure x y) = bipure x y

t (x <<*>> y) = t x <<*>> t y

Performance note

traverseBia is fairly efficient, and uses compiler rewrite rules to be even more efficient for a few important types like []. However, if performance is critical, you might consider writing a container-specific implementation.

sequenceBia :: (Traversable t, Biapplicative p) => t (p b c) -> p (t b) (t c)

Perform all the Biappicative actions in a Traversable container and produce a container with all the results.

sequenceBia = traverseBia id

traverseBiaWith :: forall p a b c s t. Biapplicative p => (forall f x. Applicative f => (a -> f x) -> s -> f (t x)) -> (a -> p b c) -> s -> p (t b) (t c)

A version of traverseBia that doesn't care how the traversal is done.

traverseBia = traverseBiaWith traverse

module Data.Bifunctor